# Conflicts:
#	src/main/resources/db/changelog/db.changelog-master.yaml

+++ src/main/java/org/example/dto/wine/WineRequestDto.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+
+import java.math.BigDecimal;
 import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
@@ -18,14 +20,14 @@ public class WineRequestDto {
     private String wineName;
     @Positive(message = "{validation.price.invalid}")
     @NotNull
-    private Double price;
+    private BigDecimal price;
     @NotBlank
     private String countryOfOrigin;
     @NotBlank
     private String wineType;
     @NotNull
     @Min(value = 0, message = "{validation.popularity-rating.invalid}")
-    private Double popularityRating;
+    private BigDecimal popularityRating;
     @NotBlank
     private List<String> occasions;
     @NotNull
